+++ src/main/Recorder.ts
@@ -1444,7 +1444,8 @@ export default class Recorder extends EventEmitter {
       item.name.startsWith('[Wow.exe]: ') ||
       item.name.startsWith('[WowT.exe]: ') ||
       item.name.startsWith('[WowB.exe]: ') ||
-      item.name.startsWith('[WowClassic.exe]: ')
+      item.name.startsWith('[WowClassic.exe]: ') ||
+      item.name.startsWith('[WowClassicT.exe]: ')
     );
   }