+++ src/main/preload.ts
@@ -1,6 +1,7 @@
 import { contextBridge, ipcRenderer, IpcRendererEvent } from 'electron';
 import { ObsProperty, SceneItemPosition, SourceDimensions } from 'noobs';
-import { AudioSourceType, SceneItem } from './types';
+import { AudioSourceType, RendererVideo, SceneItem } from './types';
+import { TChatMessage } from 'types/api';
 
 export type Channels =
   | 'window'
@@ -200,6 +201,18 @@ contextBridge.exposeInMainWorld('electron', {
 
     refreshCloudGuilds() {
       ipcRenderer.send('refreshCloudGuilds');
+    },
+
+    getOrCreateChatCorrelator(video: RendererVideo): Promise<string> {
+      return ipcRenderer.invoke('getOrCreateChatCorrelator', video);
+    },
+
+    getChatMessages(correlator: string): Promise<TChatMessage[]> {
+      return ipcRenderer.invoke('getChatMessages', correlator);
+    },
+
+    postChatMessage(correlator: string, message: string) {
+      ipcRenderer.send('postChatMessage', correlator, message);
     },
   },
 });