+++ src/renderer/CategoryPage.tsx
@@ -9,13 +9,16 @@ import {
   useRef,
 } from 'react';
 import {
-  Eye,
   GripHorizontal,
   LockKeyhole,
   Trash,
   LockOpen,
   CloudUpload,
   CloudDownload,
+  ArrowLeftFromLine,
+  ArrowRightToLine,
+  Cloud,
+  Clapperboard,
 } from 'lucide-react';
 import { getLocalePhrase } from 'localisation/translations';
 import { VideoCategory } from '../types/VideoCategory';
@@ -35,10 +38,8 @@ import MultiPovPlaybackToggles from './MultiPovPlaybackToggles';
 import VideoFilter from './VideoFilter';
 import { Resizable, ResizeCallback } from 're-resizable';
 import { Direction } from 're-resizable/lib/resizer';
-import VideoPlayer from './VideoPlayer';
+import VideoPlayer, { VideoPlayerRef } from './VideoPlayer';
 import Label from './components/Label/Label';
-import { Popover, PopoverContent } from './components/Popover/Popover';
-import { PopoverTrigger } from '@radix-ui/react-popover';
 import ViewpointSelection from './components/Viewpoints/ViewpointSelection';
 import useTable from './components/Tables/TableData';
 import { Tooltip } from './components/Tooltip/Tooltip';
@@ -47,6 +48,8 @@ import StorageFilterToggle from './StorageFilterToggle';
 import VideoCorrelator from './VideoCorrelator';
 import { Phrase } from 'localisation/phrases';
 import BulkTransferDialog from './BulkTransferDialog';
+import VideoChat from './VideoChat';
+import ConfirmChatNamePrompt from './ConfirmChatNamePrompt';
 
 interface IProps {
   category: VideoCategory;
@@ -77,9 +80,9 @@ const CategoryPage = (props: IProps) => {
     videoFilterTags,
     language,
     dateRangeFilter,
-    viewpointSelectionOpen,
     cloudStatus,
     storageFilter,
+    chatOpen,
   } = appState;
 
   const { write, del } = cloudStatus;
@@ -113,6 +116,7 @@ const CategoryPage = (props: IProps) => {
 
   // Handle to reset the video player height.
   const resizableRef = useRef<Resizable>(null);
+  const videoPlayerRef = useRef<VideoPlayerRef>(null);
 
   useEffect(() => {
     const handleWindowResize = () => {
@@ -140,6 +144,45 @@ const CategoryPage = (props: IProps) => {
     };
   }, [playerHeight]);
 
+  const renderChat = (video: RendererVideo | undefined) => {
+    if (category === VideoCategory.Clips) {
+      return (
+        <div className="flex-1 flex flex-col items-center justify-center text-foreground text-sm font-bold">
+          <Clapperboard size={35} className="mb-2" />
+          {getLocalePhrase(language, Phrase.ChatForClipsComingSoon)}
+        </div>
+      );
+    }
+
+    if (!video) {
+      return (
+        <div className="flex-1 flex flex-col items-center justify-center text-foreground text-sm font-bold">
+          <Cloud size={35} className="mb-2" />
+          {getLocalePhrase(language, Phrase.ChatUploadToCloudText)}
+        </div>
+      );
+    }
+
+    if (config.cloudAccountName !== config.chatUserNameAgreed) {
+      return (
+        <ConfirmChatNamePrompt
+          cloudAccountName={config.cloudAccountName}
+          setConfig={setConfig}
+          language={language}
+        />
+      );
+    }
+
+    return (
+      <VideoChat
+        key={video.videoName}
+        videoPlayerRef={videoPlayerRef}
+        video={video}
+        language={language}
+      />
+    );
+  };
+
   /**
    * Handle a resize event.
    */
@@ -150,6 +193,63 @@ const CategoryPage = (props: IProps) => {
   ) => {
     const height = element.clientHeight;
     playerHeight.current = height;
+  };
+
+  const renderDrawerOpen = () => {
+    // Only the first row in the selection is relevant for the drawer display.
+    const selectedRows = table.getSelectedRowModel().rows;
+    const selectedRow = selectedRows[0];
+
+    const activeParentVideo = selectedRow
+      ? selectedRow.original
+      : filteredState[0];
+
+    // Only try to find a chat video if we have a video with cloud storage,
+    // a start time and a hash, else we cannot find the chat correlator.
+    const chatVideo = [activeParentVideo, ...activeParentVideo.multiPov].find(
+      (rv) => rv.cloud && rv.uniqueHash && rv.start,
+    );
+
+    return (
+      <div className="max-w-[500px] min-w-[500px] h-full bg-background-higher flex flex-col mx-2 gap-y-2">
+        <div className="flex items-start">
+          <Button
+            onClick={() =>
+              setAppState((prev) => ({ ...prev, chatOpen: false }))
+            }
+            variant="ghost"
+            className="mt-2"
+            size="xs"
+          >
+            <ArrowRightToLine size={18} />
+          </Button>
+        </div>
+        <div className="flex items-center justify-center w-full">
+          <ViewpointSelection
+            video={activeParentVideo}
+            appState={appState}
+            setAppState={setAppState}
+            persistentProgress={persistentProgress}
+          />
+        </div>
+        {renderChat(chatVideo)}
+      </div>
+    );
+  };
+
+  const renderDrawerClosed = () => {
+    return (
+      <div className="h-full relative">
+        <Button
+          onClick={() => setAppState((prev) => ({ ...prev, chatOpen: true }))}
+          variant="ghost"
+          className="absolute top-0 right-0 m-2"
+          size="xs"
+        >
+          <ArrowLeftFromLine size={18} />
+        </Button>
+      </div>
+    );
   };
 
   /**
@@ -176,6 +276,7 @@ const CategoryPage = (props: IProps) => {
         enable={{ bottom: true }}
         bounds="parent"
         onResize={onResize}
+        minHeight={chatOpen ? 500 : undefined}
         handleStyles={{
           bottom: {
             width: '50%',
@@ -191,15 +292,21 @@ const CategoryPage = (props: IProps) => {
           bottom: <GripHorizontal />,
         }}
       >
-        <VideoPlayer
-          key={videosToPlay.map((rv) => rv.videoName + rv.cloud).join(', ')}
-          videos={videosToPlay}
-          categoryState={categoryState}
-          persistentProgress={persistentProgress}
-          config={config}
-          appState={appState}
-          setAppState={setAppState}
-        />
+        <div className="flex h-full w-full">
+          <VideoPlayer
+            ref={videoPlayerRef}
+            key={videosToPlay.map((rv) => rv.videoName + rv.cloud).join(', ')}
+            videos={videosToPlay}
+            categoryState={categoryState}
+            persistentProgress={persistentProgress}
+            config={config}
+            appState={appState}
+            setAppState={setAppState}
+          />
+
+          {chatOpen && renderDrawerOpen()}
+          {!chatOpen && renderDrawerClosed()}
+        </div>
       </Resizable>
     );
   };
@@ -427,65 +534,6 @@ const CategoryPage = (props: IProps) => {
       return <Label>{text}</Label>;
     };
 
-    const renderViewpointSelectionPopover = () => {
-      return (
-        <Popover
-          open={viewpointSelectionOpen}
-          onOpenChange={() => {
-            setAppState((a) => {
-              return {
-                ...a,
-                viewpointSelectionOpen: !a.viewpointSelectionOpen,
-              };
-            });
-          }}
-        >
-          <PopoverTrigger asChild className="absolute top-[90px] left-0">
-            <Button
-              variant={viewpointSelectionOpen ? 'default' : 'secondary'}
-              size="xs"
-              className="z-10 h-20 rounded-l-none flex justify-start p-[3px]"
-            >
-              <Eye size={15} />
-            </Button>
-          </PopoverTrigger>
-          <PopoverContent
-            onEscapeKeyDown={(e) => {
-              // Close the popover when escape is pressed.
-              setAppState((a) => {
-                return {
-                  ...a,
-                  viewpointSelectionOpen: false,
-                };
-              });
-
-              // Need this for some reason else the popover doesn't close.
-              e.preventDefault();
-              e.stopPropagation();
-            }}
-            side="left"
-            className="p-0 absolute top-[-40px] left-[30px] w-auto min-w-max"
-            onInteractOutside={(e) => {
-              e.preventDefault();
-            }}
-            onPointerDownOutside={(e) => {
-              e.preventDefault();
-            }}
-            onFocusOutside={(e) => {
-              e.preventDefault();
-            }}
-          >
-            <ViewpointSelection
-              video={selectedRow ? selectedRow.original : filteredState[0]}
-              appState={appState}
-              setAppState={setAppState}
-              persistentProgress={persistentProgress}
-            />
-          </PopoverContent>
-        </Popover>
-      );
-    };
-
     return (
       <>
         <div className="w-full flex justify-evenly items-center gap-x-2 px-4 py-2">
@@ -540,7 +588,6 @@ const CategoryPage = (props: IProps) => {
             </div>
           </div>
         </div>
-        <div className="relative">{renderViewpointSelectionPopover()}</div>
         <div className="w-full h-full overflow-hidden">
           <VideoSelectionTable
             table={table}