+++ src/renderer/preload.d.ts
@@ -1,7 +1,8 @@
 import ElectronStore from 'electron-store';
 import { Channels } from 'main/preload';
-import { AudioSourceType, SceneItem } from 'main/types';
+import { AudioSourceType, RendererVideo, SceneItem } from 'main/types';
 import { ObsProperty, SceneItemPosition, SourceDimensions } from 'noobs';
+import { TChatMessage } from 'types/api';
 
 declare global {
   interface Window {
@@ -72,6 +73,9 @@ declare global {
 
         getSensibleEncoderDefault(): Promise<string>;
         refreshCloudGuilds(): void;
+        getOrCreateChatCorrelator(video: RendererVideo): Promise<string>;
+        getChatMessages(correlator: string): Promise<TChatMessage[]>;
+        postChatMessage(correlator: string, message: string): void;
       };
     };
   }